Fixed bugs, added current_level field
Removed `insert or ignore` from the database wrappers as it risks to
surpress errors that may be important to check. Instead we need to check
insertions that already exist before adding them.

Added the current_level field as skills train slower as the number of
balls increase. May be important for prediction overall training speed.